:white_check_mark: feature tests

+++ routes/api/users.routes.php
@@ -2,25 +2,23 @@
 
 use App\Http\Controllers\UserController;
 
-Route::middleware(['auth:sanctum', 'role:admin|super_user'])
-    ->controller(UserController::class)
-    ->name('users.')
-    ->group(function () {
+Route::middleware(['auth:sanctum'])->controller(UserController::class)->name('users.')->group(function () {
         /** @uses \App\Http\Controllers\UserController::store() */
-        Route::post('', 'store')->name('store');
+        Route::middleware(['permission:create_users'])->post('', 'store')->name('store');
 
         /** @uses \App\Http\Controllers\UserController::update() */
-        Route::patch('{id}', 'update')->name('update');
+        Route::middleware(['permission:update_users'])->patch('{id}', 'update')->name('update');
 
         /** @uses \App\Http\Controllers\UserController::index() */
-        Route::get('', 'index')->name('index');
+        Route::middleware(['permission:view_users'])->get('', 'index')->name('index');
 
         /** @uses \App\Http\Controllers\UserController::read() */
-        Route::get('{id}', 'read')->name('read');
+        Route::middleware(['permission:view_users'])->get('{id}', 'read')->name('read');
 
         /** @uses \App\Http\Controllers\UserController::destroy() */
-        Route::delete('{id}', 'destroy')->name('destroy');
+        Route::middleware(['permission:delete_users'])->delete('{id}', 'destroy')->name('destroy');
 
         /** @uses \App\Http\Controllers\UserController::uploadProfilePicture() */
-        Route::post('{id}/profile-picture', 'uploadProfilePicture')->name('upload.profile-picture');
+        Route::middleware(['permission:update_users'])->post('{id}/profile-picture', 'uploadProfilePicture')
+            ->name('upload.profile-picture');
     });
